tag_to_header.py: Always write trailing newlines
Leaving off a final newline in text files is very unconventional for
Unix and breaks a number of assumptions that other common tools make.
It's also simpler to always write a newline than to keep track of
whether you're writing the first line or not.

For example, obtaining a line count with `wc -l` on the produced FastQ
files will be off by one before this change.